created some components, pulled code out of app.js
