+++ client/src/components/BasicComponent.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const BasicComponent = () => {
+  return (
+    <div>
+      <h1>Hello, this is a Basic Component!</h1>
+    </div>
+  );
+};
+
+export default BasicComponent;
